Updating the deletePet function with authorization check to see if the user is allowed to delete a pet

+++ src/actions/actions.ts
@@ -114,13 +114,28 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 }
 
 export async function deletePet(petId: unknown) {
-  const validatedId = petIdSchema.safeParse(petId);
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
 
+  const validatedId = petIdSchema.safeParse(petId);
   if (!validatedId.success) {
     return {
       message: "Invalid pet data",
     };
   }
+
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedId.data,
+    },
+  });
+  if (!pet) {
+    return { message: "Pet not found" };
+  }
+
+  if (pet.userId !== session?.user?.id) return { message: "Not authorized" };
 
   try {
     await prisma.pet.delete({
